Updates type imports
