+++ packages/tabs/src/Tab.tsx
@@ -1,49 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-
-export interface TabProps {
-  /**
-   * Content that will appear as the title in the Tab list.
-   */
-  name: React.ReactNode;
-
-  /**
-   * Content that will appear inside of Tab panel.
-   */
-  children?: React.ReactNode;
-
-  /**
-   * Boolean that determines if the Tab is disabled.
-   */
-  disabled?: boolean;
-
-  /**
-   * If Tabs component is uncontrolled, this determines what Tab will be selected on first render.
-   */
-  default?: boolean;
-
-  /**
-   * Adds a className to the root element.
-   */
-  className?: string;
-
-  /**
-   * Destination when name is rendered as `a` tag.
-   */
-  href?: string;
-
-  /**
-   * Destination when name is rendered as `Link` tag.
-   */
-  to?: string;
-
-  selected?: boolean;
-
-  ariaControl?: string;
-
-  // Done in order to support any Router system, such that TabTitle component can accept any URL destination prop.
-  [key: string]: any;
-}
+import { TabProps } from './types';
 
 /**
  * # Tab
