+++ packages/tabs/src/index.ts
@@ -1,4 +1,3 @@
 export { default as Tabs } from './Tabs';
-export type { AccessibleTabsProps as TabsProps } from './Tabs';
 export { default as Tab } from './Tab';
-export type { TabProps } from './Tab';
+export type { AccessibleTabsProps as TabsProps, TabProps } from './types';
